requirements.txt
fastapi
uvicorn
pydantic
requests
google-generativeai
python-dotenv
httpx
